backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# PII detection
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.8.4

# LLM
google-genai==1.16.0

# HTTP
httpx>=0.28.1

# DB (for dashboard stats - Phase 6, install now so imports don't break)
sqlalchemy==2.0.30
aiosqlite==0.20.0

# CORS etc.
python-multipart==0.0.9